requirements.txt
fastapi 
uvicorn
PyPDF2
python-docx
passlib[bcrypt] 
python-dotenv 
python-jose 
google-generativeai
pydantic[email]
sqlalchemy 
mysqlclient 
alembic
pymysql